nhl/requirements.txt
requests
elasticsearch
python-dotenv